changes to BertSummConfig: - move from config.py to new bert_summary_config.py - add to __init__.py for gamechangerml/configs - add docstring
