+++ gamechangerml/configs/__init__.py
@@ -8,3 +8,4 @@ from .qexp_config import QexpConfig
 from .training_config import TrainingConfig
 from .d2v_config import D2VConfig
 from .validation_config import ValidationConfig
+from .bert_summary_config import BertSummConfig
