+++ gamechangerml/configs/config.py
@@ -10,34 +10,6 @@ class DefaultConfig:
     DEFAULT_FILE_PREFIX = datetime.now().strftime("%Y%m%d")
 
 
-# for Bert Extractive Summarizer (https://pypi.org/project/bert-extractive-summarizer/)
-class BertSummConfig:
-    MODEL_ARGS = {
-        "initialize": {
-            # This gets used by the hugging face bert library to load the model, you can supply a custom trained model here
-            "model": "bert-base-uncased",
-            # If you have a pre-trained model, you can add the model class here.
-            "custom_model": None,
-            # If you have a custom tokenizer, you can add the tokenizer here.
-            "custom_tokenizer": None,
-            # Needs to be negative, but allows you to pick which layer you want the embeddings to come from.
-            "hidden": -2,
-            # It can be "mean", "median", or "max". This reduces the embedding layer for pooling.
-            "reduce_option": "mean",
-        },
-        "fit": {
-            "ratio": None,  # The ratio of sentences that you want for the final summary
-            # Parameter to specify to remove sentences that are less than 40 characters
-            "min_length": 40,
-            "max_length": 600,  # Parameter to specify to remove sentences greater than the max length
-            # Number of sentences to use. Overrides ratio if supplied.
-            "num_sentences": 2,
-        },
-        "coreference": {"greedyness": 0.4},
-        "doc_limit": 100000,
-    }
-
-
 class TopicsConfig:
 
     # topic models should be in folders named gamechangerml/models/topic_model_<date>
